Convert AlbumController to generic ImageGroupController. Added /roll/** mapping to images-servlet.xml Included type in ImageGroup.equals(). Make AlbumLinkTag an ImageGroupLinkTag. ImageGroup links now must be rooted at the root of the context (/albums). Added list of groups containing the image to the image display page. Added groupsContainingImage query.
Now we can look at rolls, links of other image groups an image is in will
show up on the display page, and all we would need to do to add another
image group type would be to add the mapping in the XML file.

Added placeholder toplevel index.html.